apps/face-service/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
dlib @ https://github.com/Murtaza-Saeed/Dlib-Precompiled-Wheels-for-Python-on-Windows-x64-Easy-Installation/raw/master/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.2.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-multipart==0.0.9
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.9.0
uvicorn==0.27.1